refactor(audio): clarify transcription service location
