+++ nanobot/audio/__init__.py
@@ -1,0 +1,2 @@
+"""Shared audio service helpers."""
+
