+++ nanobot/audio/transcription.py
