+++ tests/webui/test_transcription_ws.py
@@ -75,7 +75,7 @@ async def test_webui_transcribe_audio_rejects_oversized_audio(
     config.providers.groq.api_key = "gsk-test"
     save_config(config, config_path)
     monkeypatch.setattr("nanobot.config.loader._current_config_path", config_path)
-    monkeypatch.setattr("nanobot.transcription.get_media_dir", lambda _channel=None: tmp_path)
+    monkeypatch.setattr("nanobot.audio.transcription.get_media_dir", lambda _channel=None: tmp_path)
 
     event, payload = await webui_transcription_event({
         "request_id": "voice-1",
@@ -101,7 +101,7 @@ async def test_webui_transcribe_audio_returns_text_and_removes_temp_file(
     save_config(config, config_path)
     monkeypatch.setattr("nanobot.config.loader._current_config_path", config_path)
     monkeypatch.setattr(
-        "nanobot.transcription.get_media_dir",
+        "nanobot.audio.transcription.get_media_dir",
         lambda _channel=None: media_dir,
     )
     captured_paths: list[Path] = []
@@ -113,7 +113,7 @@ async def test_webui_transcribe_audio_returns_text_and_removes_temp_file(
         return "hello voice"
 
     monkeypatch.setattr(
-        "nanobot.transcription.transcribe_audio_file",
+        "nanobot.audio.transcription.transcribe_audio_file",
         fake_transcribe_audio_file,
     )
